util: convert aggregate from js to ts and tweak execute-code to run async #7666

+++ src/packages/backend/execute-code.ts
@@ -6,7 +6,12 @@
 // Execute code in a subprocess.
 
 import { callback } from "awaiting";
-import { spawn } from "node:child_process";
+import LRU from "lru-cache";
+import {
+  ChildProcessWithoutNullStreams,
+  spawn,
+  SpawnOptionsWithoutStdio,
+} from "node:child_process";
 import { chmod, mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
@@ -15,21 +20,32 @@ import shellEscape from "shell-escape";
 import getLogger from "@cocalc/backend/logger";
 import { aggregate } from "@cocalc/util/aggregate";
 import { callback_opts } from "@cocalc/util/async-utils";
-import { to_json, trunc, walltime } from "@cocalc/util/misc";
+import { to_json, trunc, uuid, walltime } from "@cocalc/util/misc";
 import { envForSpawn } from "./misc";
 
-import type {
-  ExecuteCodeFunctionWithCallback,
-  ExecuteCodeOptions,
-  ExecuteCodeOptionsWithCallback,
-  ExecuteCodeOutput,
+import {
+  isExecuteCodeOptionsAsyncGet,
+  type ExecuteCodeFunctionWithCallback,
+  type ExecuteCodeOptions,
+  type ExecuteCodeOptionsAsyncGet,
+  type ExecuteCodeOptionsWithCallback,
+  type ExecuteCodeOutput,
 } from "@cocalc/util/types/execute-code";
 
 const log = getLogger("execute-code");
 
+const asyncCache = new LRU<string, ExecuteCodeOutput>({
+  max: 100,
+  ttl: 1000 * 60 * 60,
+  ttlAutopurge: true,
+  allowStale: true,
+  updateAgeOnGet: true,
+  updateAgeOnHas: true,
+});
+
 // Async/await interface to executing code.
 export async function executeCode(
-  opts: ExecuteCodeOptions,
+  opts: ExecuteCodeOptions | ExecuteCodeOptionsAsyncGet,
 ): Promise<ExecuteCodeOutput> {
   return await callback_opts(execute_code)(opts);
 }
@@ -50,8 +66,17 @@ export const execute_code: ExecuteCodeFunctionWithCallback = aggregate(
 
 // actual implementation, without the aggregate wrapper
 async function executeCodeNoAggregate(
-  opts: ExecuteCodeOptions,
+  opts: ExecuteCodeOptions | ExecuteCodeOptionsAsyncGet,
 ): Promise<ExecuteCodeOutput> {
+  if (isExecuteCodeOptionsAsyncGet(opts)) {
+    const cached = asyncCache.get(opts.async_get);
+    if (cached != null) {
+      return cached;
+    } else {
+      throw new Error(`Status or result of '${opts.async_get}' not found.`);
+    }
+  }
+
   if (opts.args == null) opts.args = [];
   if (opts.timeout == null) opts.timeout = 10;
   if (opts.ulimit_timeout == null) opts.ulimit_timeout = true;
@@ -115,7 +140,48 @@ async function executeCodeNoAggregate(
       await writeFile(tempPath, cmd);
       await chmod(tempPath, 0o700);
     }
-    return await callback(doSpawn, { ...opts, origCommand });
+
+    if (opts.async_exec) {
+      // we return an ID, the caller can then use it to query the status
+      const async_limit = 1024 * 1024; // we limit how much we keep in memory, to avoid problems
+      opts.max_output = Math.min(async_limit, opts.max_output ?? async_limit);
+      const id = uuid();
+      const start = new Date();
+      const started = {
+        stdout: `Process started running at ${start.toISOString()}`,
+        stderr: "",
+        exit_code: start.getTime(),
+        async_id: id,
+      };
+      asyncCache.set(id, started);
+
+      doSpawn({ ...opts, origCommand }, (err, result) => {
+        const started = asyncCache.get(id)?.exit_code ?? 0;
+        const info = { elapsed_s: (Date.now() - started) / 1000 };
+        if (err) {
+          asyncCache.set(id, {
+            stdout: "",
+            stderr: `${err}`,
+            exit_code: 1,
+            ...info,
+          });
+        } else if (result != null) {
+          asyncCache.set(id, { ...result, ...info });
+        } else {
+          asyncCache.set(id, {
+            stdout: "",
+            stderr: `No result`,
+            exit_code: 1,
+            ...info,
+          });
+        }
+      });
+
+      return started;
+    } else {
+      // This is the blocking variant
+      return await callback(doSpawn, { ...opts, origCommand });
+    }
   } finally {
     // clean up
     if (tempDir) {
@@ -124,7 +190,10 @@ async function executeCodeNoAggregate(
   }
 }
 
-function doSpawn(opts, cb) {
+function doSpawn(
+  opts,
+  cb: (err: string | undefined, result?: ExecuteCodeOutput) => void,
+) {
   const start_time = walltime();
 
   if (opts.verbose) {
@@ -138,7 +207,7 @@ function doSpawn(opts, cb) {
       "seconds",
     );
   }
-  const spawnOptions = {
+  const spawnOptions: SpawnOptionsWithoutStdio = {
     detached: true, // so we can kill the entire process group if it times out
     cwd: opts.path,
     ...(opts.uid ? { uid: opts.uid } : undefined),
@@ -150,8 +219,8 @@ function doSpawn(opts, cb) {
     },
   };
 
-  let r,
-    ran_code = false;
+  let r: ChildProcessWithoutNullStreams;
+  let ran_code = false;
   try {
     r = spawn(opts.command, opts.args, spawnOptions);
     if (r.stdout == null || r.stderr == null) {
@@ -215,7 +284,7 @@ function doSpawn(opts, cb) {
   });
 
   r.on("exit", (code) => {
-    exit_code = code;
+    exit_code = code != null ? code : undefined;
     finish();
   });
 
@@ -317,8 +386,10 @@ function doSpawn(opts, cb) {
         );
       }
       try {
-        killed = true;
-        process.kill(-r.pid, "SIGKILL"); // this should kill process group
+        killed = true; // we set the kill flag in any case – i.e. process will no longer exist
+        if (r.pid != null) {
+          process.kill(-r.pid, "SIGKILL"); // this should kill process group
+        }
       } catch (err) {
         // Exceptions can happen, which left uncaught messes up calling code big time.
         if (opts.verbose) {
